Add publication management
We can now add and query Keywords, the next part of the schema to implement is to be able to add and remove publications and set the keywords when a publication is made.

Also bump the requirements as we're adding a new package to deal with camelizaiton.

Approved-by: TomReichardt <thomas.reichardt@students.mq.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwlandscape-compas/-/merge_requests/9