Propagate errors from celery tasks to backend and use task chain instead of linked tasks to get accurate task result
Attempting to enhance error handling, I've added try/except blocks to both python scripts called in celery tasks (run compas and generate plots) so that execptions are propagated to backend. 

There was a bug in reporting celery task results. With the way that both tasks were linked, I could only get the result of the first task , which is running a compas job. So if the first task passed while the second (plotting) failed, I got a false success and the application doesn't recognise the error and continue waiting for the plots to be generated. 
This is now fixed by using a celery chain of tasks to be able to get the result of both tasks together. If either fails, the error is reported to frontend, then it stops waiting for the plots to be generated.

Approved-by: Lewis Lakerink <llakerink@swin.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwlandscape-compas/-/merge_requests/13