Fix sluggish form issues and not knowing when the form is processing.
The New Single Binary Form component uses a big JSON object (>1.8 MB). The object is stored in state and before this change every interaction caused the plots to reprocess this data multiple times. The result was a sluggish form or the browser freezing.

This change wraps the plots in a memo function so that the rerender occurs only when the JSON data actually changes. This greatly improves the responsiveness and speed of the site. Memo functions need to be used very carefully, but in this case, I believe it's warranted.

Approved-by: emanehab99 <eali@swin.edu.au>
Approved-by: Lewis Lakerink <llakerink@swin.edu.au>

Merged by: Asher <jasherleslie@gmail.com>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwlandscape-compas/-/merge_requests/70